requirement.txt
pillow
Flask
tensorflow 